ragrefer/requirements.txt
langchain-core>=0.1.0
langchain-community>=0.0.1
langchain-google-genai>=0.1.0
gradio>=4.0.0
pypdf>=3.17.0  # For PDF loading
python-dotenv>=1.0.0
faiss-cpu>=1.7.4